+++ packages/odata-service-writer/test/unit/index.test.ts
@@ -1284,6 +1284,43 @@ describe('update', () => {
         expect(fs.exists(join(testDir, 'webapp', 'localService', 'mainService', 'metadata.xml'))).toBe(true);
     });
 
+    it('Update an existing service without YAML updates, only update service files in localService', async () => {
+        // Write dummy YAML files that could potentially be updated
+        fs.write(join(testDir, 'ui5.yaml'), '');
+        fs.write(join(testDir, 'ui5-mock.yaml'), '');
+        fs.write(join(testDir, 'ui5-local.yaml'), '');
+        // Delete metadata and annotation file from service folder
+        fs.delete(join(testDir, 'webapp', 'localService', 'mainService', 'metadata.xml'));
+        fs.delete(join(testDir, 'webapp', 'localService', 'mainService', 'SEPMRA_PROD_MAN.xml'));
+        await update(
+            testDir,
+            {
+                name: 'mainService',
+                url: 'https://localhost',
+                path: '/sap',
+                type: ServiceType.EDMX,
+                annotations: [
+                    {
+                        technicalName: 'SEPMRA_PROD_MAN',
+                        xml: '<edmx:Edmx><?xml version="1.0" encoding="utf-8"?></edmx:Edmx>'
+                    }
+                ] as EdmxAnnotationsInfo[],
+                metadata: '<edmx:Edmx><?xml version="1.0" encoding="utf-8"?></edmx:Edmx>',
+                version: OdataVersion.v4,
+                localAnnotationsName: 'annotation'
+            },
+            fs,
+            false
+        );
+        // Check that YAML are not updated with backend and mockserver middlewares
+        expect(fs.read(join(testDir, 'ui5.yaml'))).toBe('');
+        expect(fs.read(join(testDir, 'ui5-mock.yaml'))).toBe('');
+        expect(fs.read(join(testDir, 'ui5-local.yaml'))).toBe('');
+        // Check that annotation and metadata files are created
+        expect(fs.exists(join(testDir, 'webapp', 'localService', 'mainService', 'metadata.xml'))).toBe(true);
+        expect(fs.exists(join(testDir, 'webapp', 'localService', 'mainService', 'SEPMRA_PROD_MAN.xml'))).toBe(true);
+    });
+
     it('Update an existing service with changed annotations', async () => {
         await update(
             testDir,