+++ packages/fiori-elements-writer/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sap-ux/fiori-elements-writer
 
+## 2.4.8
+
+### Patch Changes
+
+-   Updated dependencies [f9e4913]
+    -   @sap-ux/odata-service-writer@0.27.4
+
 ## 2.4.7
 
 ### Patch Changes